Started merging claimbox.tsx into claimlist.tsx
